restructure /scripts and write local storage function
