+++ src/components/interfaces/noora-chat/chat/Messages.tsx
@@ -7,7 +7,7 @@ import { InformationCircleIcon } from "@heroicons/react/outline";
 import dynamic from "next/dynamic";
 import { ACTIONS, EVENTS, STATUS } from "react-joyride";
 import SpeechSynthesizer from "../../speech/SpeechSynthesizer";
-import { messageToSpeechParams } from "../../../../scripts/util";
+import { messageToSpeechParams } from "../../../../scripts/utils/util";
 const JoyRideNoSSR = dynamic(() => import("react-joyride"), { ssr: false });
 
 export default function Messages({ history, convoState }: any) {
