+++ src/components/interfaces/noora-chat/chat/NooraChat.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useRef, useState } from "react";
 import MessageBox from "./message-box/MessageBox";
 import Header from "./Header";
 import Messages from "./Messages";
-import { messageToSpeechParams } from "../../../../scripts/util";
+import { messageToSpeechParams } from "../../../../scripts/utils/util";
 import { textToSpeech } from "../../speech/SpeechSynthesizer";
 
 export default function NooraChat({
