+++ src/scripts/utils/token_util.ts
